assignment 1
