+++ app/build.gradle.kts
@@ -4,13 +4,13 @@ plugins {
 }
 
 android {
-    namespace = "com.eightbitstechnology.profileuidesign"
+    namespace = "com.gyasilarbi.profileuidesign"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = "com.eightbitstechnology.profileuidesign"
-        minSdk = 23
-        targetSdk = 34
+        applicationId = "com.gyasilarbi.profileuidesign"
+        minSdk = 24
+        targetSdk = 33
         versionCode = 1
         versionName = "1.0"
 
@@ -37,10 +37,11 @@ android {
 
 dependencies {
 
-    implementation("androidx.core:core-ktx:1.12.0")
+    implementation("androidx.core:core-ktx:1.9.0")
     implementation("androidx.appcompat:appcompat:1.6.1")
     implementation("com.google.android.material:material:1.11.0")
     implementation("androidx.constraintlayout:constraintlayout:2.1.4")
+    implementation("com.google.android.material:material:1.11.0")
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.5")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
