+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Profile UI Design</string>
+    <string name="app_name">ProfileUIDesign</string>
 </resources>
